Finds a unique candidate: "You know that each block, row and column on a Sudoku board must contain every number between 1 and 9. Therefore, if a number, say 4, can only be put in a single cell within a block/column/row, then that number is guaranteed to fit there. "
Created 3 functions that will locate the unique candidates for the row, column, and zone, based on the current row and column index.
One function brings them all together and unions their outputs to do one final check. The return is a list of the options if there are more than one, or it is a integer value representing the unique candidate that it found.

Also, created a dictionary that's key is the zone index and it returns a tuple of the row and column indices associated with that zone.
( [row_lower_bound,... row_upper_bound], [column_lower_bound,... column_upper_bound] )