Cleaned up some documentation for a few functions.
Made the unique candidate functions sub functions of the getUniqueCandidate function.

The getXNumbers functions now all take a MODE parameter. The default MODE is to return the numbers from the domain sudoku board.

Created a function called recalculate domain which takes a cell as a parameter and looks at all the cells considered adjacent to the cell and tries to simplify their domain based on any changes it sees about the board state.
This is called after the setsinglevalue function, so that it can update the cells around the cell that was just set.