IT WORKS ON ALL PUZZLES NOW.
Created a function that is called isSolved which checks if all the cells are filled.

findSingles function tries to find if any of the cells in domain are a list of length 1. Calls set single value on those.

Created a deepcopy function that lets us easily copy puzzle objects.

Modified solve to work recursively to find an answer and return it.